+++ SonicTheHedgehog/SkillStates/SonicTheHedgehog/SuperUpgrades/SuperFollowUp.cs
@@ -1,0 +1,21 @@
+﻿using EntityStates;
+using RoR2;
+using RoR2.Audio;
+using RoR2.Skills;
+using SonicTheHedgehog.Components;
+using SonicTheHedgehog.Modules;
+using System;
+using UnityEngine;
+using UnityEngine.Networking;
+
+namespace SonicTheHedgehog.SkillStates.SuperUpgrades
+{
+    public class SuperFollowUp : FollowUp
+    {
+
+        protected override float launchPushForce
+        {
+            get { return 1000f; }
+        }
+    }
+}